Initial support of capsule settings
The capsule module serves as entry point for proxy/pulp node
configuration (we don't use the proxy module directly as we
need to make sure the certs are set properly).

This patch adds support for setting the proxy on the same machine
as the master. By default, the puppet and puppetca gets configured.
Other features can be added using `--capsule-*` configurations.

By default, it also registers the proxy into Foreman.

Also the README is update to reflect the latest changes. I've removed
the node-certs-generate and node-install stuff, as the recent
changes in Katello and the installer itself probably broke this
feature, so for now, only installing all-in-one should be working.

Therefore, also the `node-certs-generate` and `node-install` scripts
are disabled.