Fixes #5012: Postgresql role already exists errors.
Problem appears to have been related to use of puppet ~> notification arrow,
which seems to cause the postgresql module to attempt to re-create the roles
again. In this case we do not appear to want or need notifications sent if a
chained task makes any changes (which is almost certainly will running this
installer the first time). Switching to the simpler -> chaining arrow corrects
the problem and maintains dependency ordering.