Fixes #6278 : Add additional checks to the installer.
This adds two new checks. One is using the output of java --version
to ensure the version is at least 1.7 and is the OpenJDK. The second
is looking for total available disk space of at least 5 GB. This does
not really ensure a space issue, but it is is a good first cut.