refs #7264 - check for java, unless installing for capsule & update elasticsearch
This commit moves the java check from a check to a hook, so that
we can more easily differentiate the behavior for a capsule vs
server install.  In addition, if the installer is being run for
the capsule, skip the java check, since the capsule does not require
it.

This commit is also updating the elasticsearch module to address
the second item raised by issue 7264.