fixes #13892 - update the README to reflect the scenario-based installer
Several changes were made to enable installing katello and the capsule
as 'scenarios' using the foreman-installer.

This commit contains changes to the katello-installer to reflect the
updated behavior, which primarily affect the commands to install
the katello and capsule (i.e. foreman-installer --scenario [katello|capsule])
and changes to the input parameters as we also separated the
puppet-capsule from foreman-proxy.