Fixes #15277 - use `yum erase` instead of `rpm -e`.
The gutterball removal requests the removal of a handful of packages.
If any are missing with `rpm -e`, the entire command will fail.
However, if you use `yum erase`, the missing packages will not cause an
issue and the remaining packages will be removed. Additionally, `yum
erase` will return a zero exit code in this case.

Logic was also added to cover both the `tomcat` and `tomcat6` dirs
during cleanup.