fixes #15350: add migrations for capsule -> foreman_proxy move
With the Katello on Foreman feature, there was an effort to separate
the foreman_proxy aspects from the puppet-capsule.  The changes
for this were primarily done as part of:
	https://github.com/Katello/puppet-capsule/pull/64

At that time, there were no migration capabilities to move the
answers from the 'capsule' to 'foreman_proxy' within the various
answer files (e.g. answers.katello-installer.yaml &
answers.capsule-installer.yaml).

This commit will add in migrations for each env to support moving
the parameters over.  This includes:
- katello
- katello-devel
- capsule

It should be noted that for katello and katello-devel include
an additional migration.  It is one that was added in a previous
commit for capsule, but it should apply to all environments.

(cherry picked from commit 04290b8e8f8892aa3723fb94284f76a5efd79d06)