Fixes #17975 - Suggest a valid variable on installer output
FOREMAN-PROXY isn't a valid identifier in bash because of the dash,
so if an user copies and pastes that command, it will never work as that
variable cannot be set. It should be something like FOREMANPROXY
or FOREMAN_PROXY so users can set that variable.

(cherry picked from commit e3136e622beba0325c1b0ed98dc6abcbc905ac27)