Fixes #17191 - don't continue with upgrade on db migrate error
Before this patch, we were ignoring the results of db:migrate rake
task in pre-upgrade hook. While sometimes it helped us with issues,
when the second run of db:migrate inside the installer would finish
successfully (just because of some combination plugins or jumping
though multiple versions). We should however rather be aware of the
issues and fixing them, rather than ignoring it, with potential
consequences later.