fixes #18405 - modify upgrade to include tasks to fix mismatches
There exist several rake tasks in katello to resolve data integrity
issues (or mismatches) between it and the backend services
(pulp and candlepin).  This commit adds those steps to the installer
upgrade process so that the issues can be addressed automatically
versus the current behavior which may result in a failed upgrade
and then a need to run them manually.

Note: this is primarily for differences related to repository
and consumer resources.

The commit also includes an option (--disable-resolve-mismatches)
to run the upgrade without making any changes.  It will still
run the tasks and report the changes that it would have made.