Changed error reporting in `check-cert-encoding`
Changed error reporting so that the `check-cert-encoding` function
output its own error message, instead of using the `error` function.

I did this so that the error could be reported to the user, and the
script then exit. Athough this would prevent the further tests being
run, there was no benefit in proceeding with those tests.
Proeeding to run the further tests, against a certificate with the wrong
encoding, would result in numerous OpenSSL error messages, which would
only confuse, rather than help, the user.