Fixes #20054 - Check server cert's encoding is PEM
Changed error reporting so that the `check-cert-encoding` function
output its own error message, instead of using the `error` function.

I did this so that the error could be reported to the user, then the
script file could exit. If the certificate's encoding is not PEM, the
subsequent tests will fail with non-obvious error messages.